refactor(account-detail): Reworked queries to start from chest table
