+++ apps/website/src/app/accounts/[tag]/(overview)/page.tsx
@@ -20,7 +20,7 @@ export default async function Page({
   const chest = await getLatestChest(tag);
 
   return (
-    <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
+    <div className="grid grid-cols-1 lg:grid-cols-3 gap-4">
       <Card className="shadow-md">
         <CardHeader>
           <CardDescription className="flex justify-between items-center">
